backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.5

# Production dependencies
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
dj-database-url==2.3.0
python-dotenv==1.0.1
